De zoomcamp tutorial

+++ 04-analytics-engineering/taxi_rides_ny/models/staging/stg_green_tripdata.sql
@@ -10,10 +10,10 @@ with tripdata as
 select
     -- identifiers
     {{ dbt_utils.surrogate_key(['vendorid', 'lpep_pickup_datetime']) }} as tripid,
-    cast(vendorid as integer) as vendorid,
-    cast(ratecodeid as integer) as ratecodeid,
-    cast(pulocationid as integer) as  pickup_locationid,
-    cast(dolocationid as integer) as dropoff_locationid,
+    cast(vendorid as numeric) as vendorid,
+    cast(ratecodeid as numeric) as ratecodeid,
+    cast(pulocationid as numeric) as  pickup_locationid,
+    cast(dolocationid as numeric) as dropoff_locationid,
     
     -- timestamps
     cast(lpep_pickup_datetime as timestamp) as pickup_datetime,
@@ -21,9 +21,9 @@ select
     
     -- trip info
     store_and_fwd_flag,
-    cast(passenger_count as integer) as passenger_count,
+    cast(passenger_count as numeric) as passenger_count,
     cast(trip_distance as numeric) as trip_distance,
-    cast(trip_type as integer) as trip_type,
+    cast(trip_type as numeric) as trip_type,
     
     -- payment info
     cast(fare_amount as numeric) as fare_amount,
@@ -34,7 +34,7 @@ select
     cast(ehail_fee as numeric) as ehail_fee,
     cast(improvement_surcharge as numeric) as improvement_surcharge,
     cast(total_amount as numeric) as total_amount,
-    coalesce(cast(payment_type as integer),0) as payment_type,
+    coalesce(cast(payment_type as numeric),0) as payment_type,
     {{ get_payment_type_description('payment_type') }} as payment_type_description
 from tripdata
 where rn = 1
